+++ src/fsCore/Common/Models/Validators/BaseValidator.cs
@@ -7,11 +7,11 @@ namespace Common.Models.Validators
     {
 
         protected static bool GuidNotNullOrEmpty(Guid? guid) => guid.HasValue && guid != Guid.Empty;
-        protected static bool LatWithInRange(double lat)
+        protected static bool LatWithInRange(decimal lat)
         {
             return lat >= -90 && lat <= 90;
         }
-        protected static bool LngWithInRange(double lng)
+        protected static bool LngWithInRange(decimal lng)
         {
             return lng >= -180 && lng <= 180;
         }
@@ -22,7 +22,7 @@ namespace Common.Models.Validators
         protected static bool NotJustBeWhiteSpace(string? input) => string.IsNullOrEmpty(input) ? true : !input.All(char.IsWhiteSpace);
         protected static bool NotJustHaveNumbers(string? input) => string.IsNullOrEmpty(input) ? true : !input.All(char.IsDigit);
         protected static bool LettersAndWhiteSpaceOnly(string input) => input.Any(char.IsLetter) && !input.All(char.IsWhiteSpace) && !input.Any(char.IsDigit) && !input.Any(char.IsPunctuation);
-        protected static bool NumbersArePositive(double input) => input > 0;
+        protected static bool NumbersArePositive(decimal input) => input > 0;
         protected static bool NumbersLettersAndWhitespaceOnlyNotJustWhiteSpaceOrNumbers(string? input)
         {
             return !string.IsNullOrEmpty(input) && input.All(x => x == char.Parse(" ") || char.IsLetterOrDigit(x)) && !input.All(char.IsWhiteSpace) && !input.All(char.IsDigit);
